# openapi.yml
openapi: 3.0.3
info:
  title: gugu-api
  version: 1.0.0
  description: gugu API의 상태 확인 및 인증 관련 엔드포인트 명세
servers:
  - url: https://price-eye.duckdns.org
    description: Production
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /v1/tracked-items/:
    get:
      tags:
        - 유저 저장 상품
      summary: Tracked Items 목록 조회
      description: 특정 유저가 대시보드에 저장한 활성 상품 목록을 조회한다. 카드의 `title`, `main_image_url`, `product_url`는 tracked item의 `currency`에 대응하는 `product_variant`를 우선 사용하고, variant가 없으면 `product` 기본값으로 fallback 한다.
      operationId: listTrackedItems
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: string
          description: 조회 대상 내부 사용자 ID
        - in: query
          name: cursor
          required: false
          schema:
            type: string
          description: 다음 페이지 커서. 첫 요청 시 생략하고, 이후 응답의 next_cursor 값을 전달한다.
        - in: query
          name: size
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: 한 번에 조회할 항목 수 (기본값 20, 최대 100)
      responses:
        '200':
          description: 조회에 성공했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTrackedItemsResponse'
        '400':
          description: 요청 파라미터가 잘못되었다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - 유저 저장 상품
      summary: Tracked Item 추가 (1~5개)
      description: 유저가 전달한 마켓과 외부 상품 식별자로 내부 상품을 조회하거나 수집한 뒤 대시보드에 저장한다. 한 번에 최대 5개까지 추가할 수 있다.
      operationId: addTrackedItems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTrackedItemRequest'
      responses:
        '201':
          description: 저장에 성공했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddTrackedItemResponse'
        '400':
          description: 요청 형식이 잘못되었거나 지원하지 않는 마켓, 상품을 찾을 수 없는 경우다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류가 발생했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tracked-items/{trackedItemID}:
    get:
      tags:
        - 유저 저장 상품
      summary: Tracked Item 상세 조회
      description: 유저 저장 상품의 상세 정보(상품 정보, SKU 목록 포함)를 조회한다. `title`, `main_image_url`, `product_url`는 tracked item의 `currency`에 대응하는 `product_variant`를 우선 사용하고, variant가 없으면 `product` 기본값으로 fallback 한다.
      operationId: getTrackedItemDetail
      parameters:
        - in: path
          name: trackedItemID
          required: true
          schema:
            type: string
          description: 유저 저장 상품 ID
        - in: query
          name: user_id
          required: true
          schema:
            type: string
          description: 조회 요청 사용자 ID
      responses:
        '200':
          description: 조회에 성공했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackedItemDetailResponse'
        '400':
          description: 요청 값이 잘못되었거나 조회 대상이 없다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류가 발생했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tracked-items/{trackedItemID}/price-alert:
    get:
      tags:
        - 유저 저장 상품
      summary: Tracked Item SKU 가격 알림 상태 조회
      description: 유저 저장 상품 문맥에서 특정 SKU의 가격 알림 상태를 조회한다. `sku_id`가 있으면 해당 SKU 기준으로 조회하고, 없으면 현재 저장된 선택 SKU 기준으로 조회한다. alert row가 없거나 선택 SKU가 없으면 `enabled=false`를 반환한다.
      operationId: getTrackedItemPriceAlert
      parameters:
        - in: path
          name: trackedItemID
          required: true
          schema:
            type: string
          description: 유저 저장 상품 ID
        - in: query
          name: sku_id
          required: false
          schema:
            type: string
          description: 조회할 SKU ID. 생략 시 tracked item에 저장된 선택 SKU 사용
        - in: query
          name: user_id
          required: true
          schema:
            type: string
          description: 조회 요청 사용자 ID
      responses:
        '200':
          description: 조회에 성공했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackedItemPriceAlertResponse'
        '400':
          description: sku_id가 현재 tracked item 상품에 속하지 않거나 요청이 잘못되었다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류가 발생했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 유저 저장 상품
      summary: Tracked Item 삭제
      description: 유저 저장 상품을 soft delete 한다.
      operationId: deleteTrackedItem
      parameters:
        - in: path
          name: trackedItemID
          required: true
          schema:
            type: string
          description: 삭제 대상 유저 저장 상품 ID
        - in: query
          name: user_id
          required: true
          schema:
            type: string
          description: 삭제 요청 사용자 ID
      responses:
        '200':
          description: 삭제에 성공했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 요청 값이 잘못되었거나 삭제 대상이 없다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류가 발생했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tracked-items/{trackedItemID}/sku:
    patch:
      tags:
        - 유저 저장 상품
      summary: Tracked Item SKU 선택
      description: 유저 저장 상품에 특정 SKU를 선택한다.
      operationId: selectTrackedItemSKU
      parameters:
        - in: path
          name: trackedItemID
          required: true
          schema:
            type: string
          description: SKU를 선택할 유저 저장 상품 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectSKURequest'
      responses:
        '200':
          description: SKU 선택에 성공했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 요청 값이 잘못되었거나 대상을 찾을 수 없다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류가 발생했다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tracked-items/{trackedItemID}/sku-price-histories:
    get:
      tags:
        - 유저 저장 상품
      summary: SKU 가격 변동 내역 조회
      description: 유저 저장 상품에서 특정 SKU의 가격 변동 내역을 조회한다. `currency` query가 있으면 해당 통화로 override 하고, 없으면 tracked item의 `currency`를 기본값으로 사용한다. 해당 통화 데이터가 없으면 빈 배열을 반환한다.
      operationId: getSKUPriceHistories
      parameters:
        - in: path
          name: trackedItemID
          required: true
          schema:
            type: string
          description: 유저 저장 상품 ID
        - in: query
          name: sku_id
          required: true
          schema:
            type: string
          description: 조회할 SKU ID
        - in: query
          name: currency
          required: false
          schema:
            type: string
          description: 통화 코드. 지정하면 tracked item 기본 통화를 override 한다. 생략 시 tracked item currency 사용.
      responses:
        '200':
          description: 조회에 성공했다. 해당 통화 이력이 없으면 빈 배열을 반환한다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SKUPriceHistoryResponse'
        '400':
          description: sku_id가 누락되었거나 요청이 잘못되었다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 유저 저장 상품을 찾을 수 없다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ResultType:
      type: string
      description: API 처리 결과 구분값
      enum:
        - SUCCESS
        - ERROR
    ErrorMessage:
      type: object
      description: 에러 응답 본문
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 애플리케이션 에러 코드
          example: E400
        message:
          type: string
          description: 사용자 또는 클라이언트가 확인할 에러 메시지
          example: invalid request body
        data:
          nullable: true
          description: 추가 에러 데이터
    ErrorResponse:
      type: object
      description: 공통 에러 응답 래퍼
      required:
        - result
        - error
      properties:
        result:
          $ref: '#/components/schemas/ResultType'
        error:
          $ref: '#/components/schemas/ErrorMessage'
...
    AddTrackedItemRequest:
      type: object
      description: 유저 저장 상품 추가 요청 (1~5개). `currency`, `language`는 프론트가 현재 화면 기준으로 함께 전달한다.
      required:
        - items
      properties:
        items:
          type: array
          description: 추가할 상품 목록 (최소 1개, 최대 5개)
          minItems: 1
          maxItems: 5
          items:
            type: object
            required:
              - original_url
              - provider_commerce
              - external_product_id
            properties:
              original_url:
                type: string
                description: 사용자가 붙여넣은 원본 URL
                example: https://www.aliexpress.com/item/1005000000000000.html
              provider_commerce:
                type: string
                description: 프론트가 판별한 마켓 식별자
                example: ALIEXPRESS
              external_product_id:
                type: string
                description: 프론트가 추출한 외부 상품 식별자
                example: '1005000000000000'
              currency:
                type: string
                description: 유저가 원하는 통화. tracked item 기본 통화와 variant 적재 기준으로 사용한다.
                example: KRW
              language:
                type: string
                description: 유저가 원하는 언어. variant 적재 및 조회 기준으로 사용한다.
                example: KO
...
    SuccessResponse:
      type: object
      description: 데이터 없이 성공만 반환하는 응답
      required:
        - result
      properties:
        result:
          $ref: '#/components/schemas/ResultType'
...
    TrackedItemData:
      type: object
      description: 유저 저장 상품 카드 데이터. `title`, `main_image_url`, `product_url`는 tracked item currency 기준 variant 우선, 없으면 product fallback 이다.
      required:
        - tracked_item_id
        - product_id
        - market
        - external_product_id
        - original_url
        - title
        - main_image_url
        - current_price
        - currency
        - product_url
      properties:
        tracked_item_id:
          type: string
          description: 유저 저장 상품 ID
        product_id:
          type: string
          description: 내부 상품 ID
        market:
          type: string
          description: 마켓 식별자
          example: ALIEXPRESS
        external_product_id:
          type: string
          description: 외부 상품 식별자
        original_url:
          type: string
          description: 사용자가 추가할 때 입력한 원본 URL
        title:
          type: string
          description: 상품명. tracked item currency 기준 variant 우선, 없으면 product 기본값 사용.
        main_image_url:
          type: string
          description: 대표 이미지 URL. tracked item currency 기준 variant 우선, 없으면 product 기본값 사용.
        current_price:
          type: string
          description: 카드에 표시할 대표 가격. tracked item currency 기준 variant의 current_price를 사용한다.
        currency:
          type: string
          description: 가격 통화
          example: KRW
        product_url:
          type: string
          description: 외부 상품 상세 URL. tracked item currency 기준 variant 우선, 없으면 product 기본값 사용.
        promotion_link:
          type: string
          nullable: true
          description: Affiliate 프로모션 링크. 빈 값이면 product_url 사용.
    TrackedItemDetailData:
      type: object
      description: 유저 저장 상품 상세 데이터. `title`, `main_image_url`, `product_url`는 tracked item currency 기준 variant 우선, 없으면 product fallback 이다.
      required:
        - tracked_item_id
        - product_id
        - market
        - external_product_id
        - original_url
        - title
        - main_image_url
        - currency
        - product_url
        - skus
      properties:
        tracked_item_id:
          type: string
          description: 유저 저장 상품 ID
        product_id:
          type: string
          description: 내부 상품 ID
        sku_id:
          type: string
          description: 선택된 SKU ID
        market:
          type: string
          description: 마켓 식별자
          example: ALIEXPRESS
        external_product_id:
          type: string
          description: 외부 상품 식별자
        original_url:
          type: string
          description: 사용자가 추가할 때 입력한 원본 URL
        title:
          type: string
          description: 상품명. tracked item currency 기준 variant 우선, 없으면 product 기본값 사용.
        main_image_url:
          type: string
          description: 대표 이미지 URL. tracked item currency 기준 variant 우선, 없으면 product 기본값 사용.
        currency:
          type: string
          description: 가격 통화
          example: KRW
        product_url:
          type: string
          description: 외부 상품 상세 URL. tracked item currency 기준 variant 우선, 없으면 product 기본값 사용.
        promotion_link:
          type: string
          nullable: true
          description: Affiliate 프로모션 링크. 빈 값이면 product_url 사용.
        skus:
          type: array
          items:
            $ref: '#/components/schemas/ProductSKUData'
    PriceAlertStateData:
      type: object
      description: 선택된 SKU 기준 가격 알림 상태. alert row가 없으면 `enabled=false` 이다.
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: 선택된 SKU의 가격 알림 활성화 여부
        channel:
          type: string
          description: 알림 채널. alert row가 있을 때만 내려간다.
    TrackedItemPriceAlertResponse:
      type: object
      description: 유저 저장 상품 문맥의 SKU 가격 알림 상태 응답
      required:
        - result
        - data
      properties:
        result:
          $ref: '#/components/schemas/ResultType'
        data:
          $ref: '#/components/schemas/PriceAlertStateData'
    TrackedItemDetailResponse:
      type: object
      description: 유저 저장 상품 상세 응답
      required:
        - result
        - data
      properties:
        result:
          $ref: '#/components/schemas/ResultType'
        data:
          $ref: '#/components/schemas/TrackedItemDetailData'
    AddTrackedItemEntry:
      type: object
      description: 개별 상품 추가 결과
      required:
        - tracked_item_id
        - already_tracked
      properties:
        tracked_item_id:
          type: string
          description: 유저 저장 상품 ID
        already_tracked:
          type: boolean
          description: 이미 저장된 상품이어서 기존 관계를 재사용했는지 여부
    AddTrackedItemData:
      type: object
      description: 유저 저장 상품 추가 응답 데이터
      required:
        - results
        - total
        - added
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/AddTrackedItemEntry'
        total:
          type: integer
          description: 요청한 상품 수
          example: 3
        added:
          type: integer
          description: 새로 추가된 상품 수 (already_tracked 제외)
          example: 2
    AddTrackedItemResponse:
      type: object
      description: 유저 저장 상품 추가 응답
      required:
        - result
        - data
      properties:
        result:
          $ref: '#/components/schemas/ResultType'
        data:
          $ref: '#/components/schemas/AddTrackedItemData'
    ListTrackedItemsResponse:
      type: object
      description: 유저 저장 상품 목록 응답 (커서 기반 페이지네이션)
      required:
        - result
        - data
      properties:
        result:
          $ref: '#/components/schemas/ResultType'
        data:
          $ref: '#/components/schemas/ListTrackedItemsPageData'
    ListTrackedItemsPageData:
      type: object
      description: 커서 기반 페이지네이션 목록 데이터
      required:
        - items
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TrackedItemData'
        next_cursor:
          type: string
          description: 다음 페이지를 조회하기 위한 커서 값. has_more가 false이면 생략된다.
        has_more:
          type: boolean
          description: 다음 페이지 존재 여부
    SKUPriceHistoryResponse:
      type: object
      description: SKU 가격 변동 내역 응답. 요청 통화 이력이 없으면 `data`는 빈 배열이다.
      required:
        - result
        - data
      properties:
        result:
          $ref: '#/components/schemas/ResultType'
        data:
          type: array
          items:
            $ref: '#/components/schemas/SKUPriceHistoryItem'
    SKUPriceHistoryItem:
      type: object
      description: SKU 가격 변동 내역 항목
      required:
        - sku_id
        - price
        - currency
        - recorded_at
        - change_value
      properties:
        sku_id:
          type: string
          description: SKU ID
        price:
          type: string
          description: 기록 시점 가격
        currency:
          type: string
          description: 가격 통화
        recorded_at:
          type: string
          format: date-time
          description: 기록 시각
        change_value:
          type: string
          description: 이전 기록 대비 변동값
...
    SelectSKURequest:
      type: object
      description: Tracked Item SKU 선택 요청
      required:
        - user_id
        - sku_id
      properties:
        user_id:
          type: string
          description: 내부 사용자 ID
          example: user_01
        sku_id:
          type: string
          description: 선택할 SKU ID
          example: sku_01
    ProductSKUData:
      type: object
      description: 상품 SKU(옵션) 데이터
      required:
        - id
        - external_sku_id
        - sku_name
        - color
        - size
        - price
        - original_price
        - currency
        - image_url
      properties:
        id:
          type: string
          description: SKU 고유 ID
        external_sku_id:
          type: string
          description: 외부 SKU 식별자
        sku_name:
          type: string
          description: SKU 이름
        color:
          type: string
          description: 색상
        size:
          type: string
          description: 사이즈
        price:
          type: string
          description: 현재 가격
        original_price:
          type: string
          description: 원래 가격
        currency:
          type: string
          description: 가격 통화
          example: KRW
        image_url:
          type: string
          description: SKU 이미지 URL
        origin_sku_id:
          type: string
          description: 외부 마켓 원본 SKU ID
        sku_properties:
          type: string
          description: 'SKU 속성 JSON 문자열 예: [{"색상":"000000","크기":"XXXL"}]'
...